requirements.txt
groq==0.5.0
streamlit==1.32.0
python-dotenv==1.0.1
tenacity==8.2.3
pytest==8.1.1
